mcp_server/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests # Assuming this is used by weather_api_tool, or remove if not
asyncio
httpx
authlib
python-jose
python-multipart
itsdangerous
